Add ability to pass middleware to getting data from api
