+++ workers/loc.api/queue/write-data-to-stream/index.js
@@ -2,10 +2,10 @@
 
 const { cloneDeep } = require('lodash')
 
+const { getDataFromApi } = require('../../helpers')
 const {
   writeMessageToStream,
   setDefaultPrams,
-  getDataFromApi,
   filterMovementsByAmount,
   write,
   progress
